+++ lib/index.js
@@ -56,7 +56,7 @@ function isSubset(a, b) {
  *   use in generating spelling and stopwords dictionaries for search
  * @param {Object} options [{}] - A dictionary with options. For a
  * list of available options see {@link
- * configparse_options|config.parse_options}
+ * #configparseoptions|config.parseOptions}
 */
 class Hatch {
     constructor(name, language, options) {
